+++ packages/@justaname.id/sdk/src/lib/api/axiosController.ts
@@ -11,7 +11,8 @@ function loadEnv() {
 
   if (typeof process !== 'undefined' && process.env) {
     try {
-      isDevelopment = process.env['JUSTANAME_ENVIRONMENT'] === 'development' ||
+      isDevelopment =
+        process.env['JUSTANAME_ENVIRONMENT'] === 'development' ||
         process.env['NEXT_PUBLIC_JUSTANAME_ENVIRONMENT'] === 'development' ||
         process.env['VITE_JUSTANAME_ENVIRONMENT'] === 'development' ||
         process.env['NODE_ENV'] === 'test';
@@ -20,7 +21,9 @@ function loadEnv() {
     }
   }
 
-  BASE_URL = isDevelopment ? 'https://api-staging.justaname.id' : 'https://api.justaname.id';
+  BASE_URL = isDevelopment
+    ? 'https://api-staging.justaname.id'
+    : 'https://api.justaname.id';
 }
 
 loadEnv();
@@ -35,14 +38,13 @@ export const justANameInstance = axios.create({
   baseURL: getBaseUrl(),
 });
 
-
 /**
  * Represents the Controlled Axios Promise type.
  * @typeparam T - The type of the data to be returned.
  */
-export type ControlledAxiosPromise<T extends {}> =
-  AxiosPromise<BaseResponse<T>>;
-
+export type ControlledAxiosPromise<T extends {}> = AxiosPromise<
+  BaseResponse<T>
+>;
 
 /**
  * Represents the controlled axios promise.